requirements.txt
# Core dependencies
fastapi==0.100.0
uvicorn[standard]==0.23.0
httpx==0.24.0
python-dotenv==1.0.0
pydantic==1.10.7
pytest==7.4.0

# Database
aiosqlite==0.19.0

# Optional UI
streamlit==1.30.0

# Optional: Uncomment for OpenAI integration
# openai==1.0.0

# Optional: Uncomment for AWS Bedrock integration
# boto3==1.28.0